changes structure of IOS plugin slightly
