+++ ios/Classes/IosARView.swift
@@ -1,29 +1,10 @@
 import Flutter
 import UIKit
-
-class IosARViewFactory: NSObject, FlutterPlatformViewFactory {
-    private var messenger: FlutterBinaryMessenger
-
-    init(messenger: FlutterBinaryMessenger) {
-        self.messenger = messenger
-        super.init()
-    }
-
-    func create(
-        withFrame frame: CGRect,
-        viewIdentifier viewId: Int64,
-        arguments args: Any?
-    ) -> FlutterPlatformView {
-        return IosARView(
-            frame: frame,
-            viewIdentifier: viewId,
-            arguments: args,
-            binaryMessenger: messenger)
-    }
-}
+import Foundation
+import ARKit
 
 class IosARView: NSObject, FlutterPlatformView {
-    private var _view: UIView
+    private var _sceneView: ARSCNView
 
     init(
         frame: CGRect,
@@ -31,23 +12,23 @@ class IosARView: NSObject, FlutterPlatformView {
         arguments args: Any?,
         binaryMessenger messenger: FlutterBinaryMessenger?
     ) {
-        _view = UIView()
+        _sceneView = ARSCNView(frame: frame)
         super.init()
-        // iOS views can be created here
-        createNativeView(view: _view)
+        _sceneView.delegate = self
+        _sceneView.seesion.run()
     }
 
     func view() -> UIView {
-        return _view
+        return _sceneView
     }
 
-    func createNativeView(view _view: UIView){
-        _view.backgroundColor = UIColor.blue
-        let nativeLabel = UILabel()
-        nativeLabel.text = "Native text from iOS"
-        nativeLabel.textColor = UIColor.white
-        nativeLabel.textAlignment = .center
-        nativeLabel.frame = CGRect(x: 0, y: 0, width: 180, height: 48.0)
-        _view.addSubview(nativeLabel)
-    }
+    //func createNativeView(view _view: UIView){
+    //    _view.backgroundColor = UIColor.blue
+    //    let nativeLabel = UILabel()
+    //    nativeLabel.text = "Native text from iOS"
+    //    nativeLabel.textColor = UIColor.white
+    //    nativeLabel.textAlignment = .center
+    //    nativeLabel.frame = CGRect(x: 0, y: 0, width: 180, height: 48.0)
+    //    _view.addSubview(nativeLabel)
+    //}
 }
